+++ tests/integration/test_scripting_engine_popups.gd
@@ -58,7 +58,7 @@ func test_filtered_multiple_choice():
 	yield(yield_for(0.1), YIELD)
 	menu = board.get_node("CardChoices")
 	assert_null(menu, "menu should not appear when filter does not match")
-
+#
 func test_task_confirm_dialog() -> void:
 	card.scripts = {"manual": {"hand": [
 				{"name": "flip_card",
@@ -67,10 +67,12 @@ func test_task_confirm_dialog() -> void:
 				"set_faceup": false}]}}
 	card.execute_scripts()
 	var confirm = board.get_node("OptionalConfirmation")
-	assert_true(confirm.visible)
-	confirm._on_OptionalConfirmation_confirmed()
-	assert_true(confirm.is_accepted, "Confirmation dialog accepted")
-	confirm.hide()
+	assert_not_null(confirm)
+	if confirm:
+		assert_true(confirm.visible)
+		confirm._on_OptionalConfirmation_confirmed()
+		assert_true(confirm.is_accepted, "Confirmation dialog accepted")
+		confirm.hide()
 	yield(yield_to(card._flip_tween, "tween_all_completed", 0.5), YIELD)
 	assert_false(card.is_faceup,
 			"Card should be face-down after accepted dialog")
@@ -87,9 +89,11 @@ func test_task_confirm_dialog() -> void:
 	assert_true(target.is_faceup,
 			"Card should not be face-down until after accepted dialog")
 	confirm = board.get_node("OptionalConfirmation")
-	confirm._on_OptionalConfirmation_cancelled()
-	assert_false(confirm.is_accepted, "Confirmation dialog not accepted")
-	confirm.hide()
+	assert_not_null(confirm)
+	if confirm:
+		confirm._on_OptionalConfirmation_cancelled()
+		assert_false(confirm.is_accepted, "Confirmation dialog not accepted")
+		confirm.hide()
 	yield(yield_to(target._flip_tween, "tween_all_completed", 0.5), YIELD)
 	assert_false(target.is_faceup,
 			"Card should be face-down after even afer other optional task cancelled")
@@ -110,8 +114,10 @@ func test_task_confirm_cost_dialog_cancelled_target() -> void:
 			"dest_container":  cfc.NMAP.discard}]}}
 	card.execute_scripts()
 	confirm = board.get_node("OptionalConfirmation")
-	confirm._on_OptionalConfirmation_cancelled()
-	confirm.hide()
+	assert_not_null(confirm)
+	if confirm:
+		confirm._on_OptionalConfirmation_cancelled()
+		confirm.hide()
 	yield(yield_to(card._flip_tween, "tween_all_completed", 0.5), YIELD)
 	assert_true(card.is_faceup,
 			"Card should not be face-down with a cancelled cost dialog")
@@ -132,8 +138,10 @@ func test_task_confirm_cost_dialogue_accepted_target() -> void:
 			"dest_container":  cfc.NMAP.discard}]}}
 	card.execute_scripts()
 	confirm = board.get_node("OptionalConfirmation")
-	confirm._on_OptionalConfirmation_confirmed()
-	confirm.hide()
+	assert_not_null(confirm)
+	if confirm:
+		confirm._on_OptionalConfirmation_confirmed()
+		confirm.hide()
 	yield(yield_for(0.5), YIELD)
 	assert_true(card.targeting_arrow.is_targeting,
 			"Card started targeting once dialogue accepted")
@@ -157,8 +165,10 @@ func test_script_confirm_dialog() -> void:
 	yield(table_move(card, Vector2(500,200)), "completed")
 	card.execute_scripts()
 	confirm = board.get_node("OptionalConfirmation")
-	confirm._on_OptionalConfirmation_cancelled()
-	confirm.hide()
+	assert_not_null(confirm)
+	if confirm:
+		confirm._on_OptionalConfirmation_cancelled()
+		confirm.hide()
 	yield(yield_to(card._flip_tween, "tween_all_completed", 0.5), YIELD)
 	assert_true(card.is_faceup,
 			"Card has not have executed any tasks with cancelled script dialog")
@@ -166,8 +176,10 @@ func test_script_confirm_dialog() -> void:
 			"Card has not have executed any tasks with cancelled script dialog")
 	card.execute_scripts()
 	confirm = board.get_node("OptionalConfirmation")
-	confirm._on_OptionalConfirmation_confirmed()
-	confirm.hide()
+	assert_not_null(confirm)
+	if confirm:
+		confirm._on_OptionalConfirmation_confirmed()
+		confirm.hide()
 	yield(yield_to(card._flip_tween, "tween_all_completed", 0.5), YIELD)
 	assert_false(card.is_faceup,
 			"Card execute all tasks properly after script confirm")