+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## 1.8 (Ongoing)
 
+### Important
+
+CFConst  has been adjusted. The following enums have been moved to CFInt, as the developers should not have a reason to modify them: `FocusStyle`, `IndexShiftPriority`, `OverlapShiftDirection`.
+You can delete these enums from your own CFConst at your leisure. However if you delete them, you need to adjust CFConst.FOCUS_STYLE
+
+
 ## New Features
 
 * 1/10 deck names with adjectives will now also contain an adverb.
@@ -11,6 +17,8 @@
 
 * Can now filter individual script tasks based on the trigger card.
 * per can now calculate based on the properties/tokens on previous subjects using `"subject": "previous"`. However for this to work, the parent task also has to have `"subject": "previous"` set.
+* Can now set "is_else" : true tasks, which will only be executed if the "is_cost" tasks cannot be paid
+* Refactored the way the script task loop is handled. If you've extended the ScriptingEngine, you will need to adjust your _init()
 
 
 ## 1.7